+++ terratex_reallife/ENVIRO/fraktionen/lspd/abschleppen.lua
@@ -222,18 +222,13 @@ function getcar_func(thePlayer, cmd, IDs)
                         vioSetElementData(thevehicle, "kmstand", dasatz["kmstand"])
                         vioSetElementData(thevehicle, "premColor", "-1")
                         local time = getRealTime()
-                        local premiumOutTime = MySql.helper.getValueSync("user_premium", "PremiumUntil", { Name = dasatz["Besitzer"] }) - time.timestamp;
 
-                        if (premiumOutTime > 0) then
-                            local lights = getStringComponents(vioGetElementData(thevehicle, "Lichterfarbe"))
-                            setVehicleHeadLightColor(thevehicle, tonumber(lights[1]), tonumber(lights[2]), tonumber(lights[3]))
-                            vioSetElementData(thevehicle, "premColor", dasatz["premColors"])
-                            if (vioGetElementData(thevehicle, "premColor") ~= "-1") then
-                                local premcolor = getStringComponents(vioGetElementData(thevehicle, "premColor"))
-                                setVehicleColor(thevehicle, tonumber(premcolor[1]), tonumber(premcolor[2]), tonumber(premcolor[3]), tonumber(premcolor[4]), tonumber(premcolor[5]), tonumber(premcolor[6]))
-                            end
-                        else
-                            vioSetElementData(thevehicle, "Lichterfarbe", "255|255|255")
+                        local lights = getStringComponents(vioGetElementData(thevehicle, "Lichterfarbe"))
+                        setVehicleHeadLightColor(thevehicle, tonumber(lights[1]), tonumber(lights[2]), tonumber(lights[3]))
+                        vioSetElementData(thevehicle, "premColor", dasatz["premColors"])
+                        if (vioGetElementData(thevehicle, "premColor") ~= "-1") then
+                            local premcolor = getStringComponents(vioGetElementData(thevehicle, "premColor"))
+                            setVehicleColor(thevehicle, tonumber(premcolor[1]), tonumber(premcolor[2]), tonumber(premcolor[3]), tonumber(premcolor[4]), tonumber(premcolor[5]), tonumber(premcolor[6]))
                         end
 
                         setVehiclePaintjob(thevehicle, tonumber(dasatz["paintjob"]))